Rename assets names in order.

+++ tx/op/offer.go
@@ -7,6 +7,7 @@ import (
 	pb "github.com/golang/protobuf/proto"
 
 	"github.com/ultiledger/go-ultiledger/account"
+	"github.com/ultiledger/go-ultiledger/crypto"
 	"github.com/ultiledger/go-ultiledger/db"
 	"github.com/ultiledger/go-ultiledger/exchange"
 	"github.com/ultiledger/go-ultiledger/ultpb"
@@ -18,8 +19,8 @@ type Offer struct {
 	AM        *account.Manager
 	EM        *exchange.Manager
 	AccountID string
-	SellAsset *ultpb.Asset
 	BuyAsset  *ultpb.Asset
+	SellAsset *ultpb.Asset
 	Amount    int64
 	Price     *ultpb.Price
 	OfferID   string
@@ -38,41 +39,59 @@ func (of *Offer) Apply(dt db.Tx) error {
 
 	sellTrust, buyTrust, err := of.loadTrust(dt)
 	if err != nil {
-		return fmt.Errorf("load selling and buying trust failed: %s", err)
+		return fmt.Errorf("load selling and buying trust failed: %v", err)
 	}
 
 	newOffer := true
 
-	if of.OfferID != "" {
+	if of.OfferID != "" { // existing offer
 		offer, err := of.EM.GetOffer(dt, of.SellAsset.AssetName, of.BuyAsset.AssetName, of.OfferID)
 		if err != nil {
 			return fmt.Errorf("get offer failed: %v", err)
 		}
+		if offer == nil {
+			return errors.New("offer not exists")
+		}
 
-		// TODO(bobonovski) release liability
-
-		err = of.EM.DeleteOffer(dt, of.SellAsset.AssetName, of.BuyAsset.AssetName, offer.OfferID)
+		// relese the liability of the offer
+		err = of.updateLiability(dt, false)
 		if err != nil {
-			return fmt.Errorf("delete offer failed: %v", err)
+			return fmt.Errorf("release offer liability failed: %v", err)
 		}
 
 		newOffer = false
 
 		// delete the offer
 		if of.Amount == 0 {
+			err = of.EM.DeleteOffer(dt, of.SellAsset.AssetName, of.BuyAsset.AssetName, offer.OfferID)
+			if err != nil {
+				return fmt.Errorf("delete offer failed: %v", err)
+			}
+
 			err = of.AM.UpdateEntryCount(acc, int32(-1))
 			if err != nil {
 				return fmt.Errorf("decrease account entry account failed: %v", err)
 			}
+
 			err = of.AM.SaveAccount(dt, acc)
 			if err != nil {
 				return fmt.Errorf("save account failed: %v", err)
 			}
 			return nil
 		}
+	} else {
+		if of.Amount == 0 {
+			return nil
+		}
+		// create a offer id for the new offer
+		offerID, err := crypto.GetOfferID()
+		if err != nil {
+			return fmt.Errorf("create offer id failed: %v", err)
+		}
+		of.OfferID = offerID
 	}
 
-	// create a new offer
+	// create a new offer with the same offer id
 	sellOffer := &ultpb.Offer{
 		AccountID: of.AccountID,
 		SellAsset: of.SellAsset,
@@ -83,15 +102,33 @@ func (of *Offer) Apply(dt db.Tx) error {
 		Passive:   of.Passive,
 	}
 
+	sellLimit, buyLimit, err := of.getOfferLimits(dt, acc, buyTrust, sellTrust, sellOffer, newOffer)
+	if err != nil {
+		return fmt.Errorf("get offer limits failed: %v", err)
+	}
+
+	if buyLimit == 0 {
+		return errors.New("offer buying limit is zero")
+	}
+	// For test
+	fmt.Println(sellLimit, buyLimit)
+
+	return nil
+}
+
+// Get the buying and selling limits of the offer.
+func (of *Offer) getOfferLimits(dt db.Tx, acc *ultpb.Account,
+	buyTrust *ultpb.Trust, sellTrust *ultpb.Trust, offer *ultpb.Offer, newOffer bool) (int64, int64, error) {
 	// temporarily increase entry count
+	var err error
 	if newOffer {
 		err = of.AM.UpdateEntryCount(acc, int32(1))
 		if err != nil {
-			return fmt.Errorf("increase account entry account failed: %v", err)
+			return -1, -1, fmt.Errorf("increase account entry account failed: %v", err)
 		}
 		err = of.AM.SaveAccount(dt, acc)
 		if err != nil {
-			return fmt.Errorf("save account failed: %v", err)
+			return -1, -1, fmt.Errorf("save account failed: %v", err)
 		}
 	}
 
@@ -103,8 +140,12 @@ func (of *Offer) Apply(dt db.Tx) error {
 		buyLimit = of.AM.GetTrustRestLimit(buyTrust)
 	}
 
-	if buyLimit < of.GetBuyLiability(sellOffer) {
-		return fmt.Errorf("account out of buying limit")
+	bl, err := of.getBuyingLiability()
+	if err != nil {
+		return -1, -1, fmt.Errorf("get buying liability failed: %v", err)
+	}
+	if buyLimit < bl {
+		return -1, -1, fmt.Errorf("out of buying limit")
 	}
 
 	// get the limit of SellAsset we can sell
@@ -115,41 +156,34 @@ func (of *Offer) Apply(dt db.Tx) error {
 		sellLimit = of.AM.GetTrustBalance(sellTrust)
 	}
 
-	if sellLimit < of.GetSellLiability(sellOffer) {
-		return fmt.Errorf("account underfund")
+	sl, err := of.getSellingLiability()
+	if err != nil {
+		return -1, -1, fmt.Errorf("get selling liability failed: %v", err)
+	}
+	if sellLimit < sl {
+		return -1, -1, fmt.Errorf("account underfund")
 	}
 
 	// decrease entry count
 	if newOffer {
 		err = of.AM.UpdateEntryCount(acc, int32(-1))
 		if err != nil {
-			return fmt.Errorf("increase account entry account failed: %v", err)
+			return -1, -1, fmt.Errorf("increase account entry account failed: %v", err)
 		}
 		err = of.AM.SaveAccount(dt, acc)
 		if err != nil {
-			return fmt.Errorf("save account failed: %v", err)
+			return -1, -1, fmt.Errorf("save account failed: %v", err)
 		}
 	}
 
 	if buyLimit == 0 {
-		return fmt.Errorf("account out of buying limit")
+		return -1, -1, fmt.Errorf("account out of buying limit")
 	}
 
 	// get the sell limit for the offer
-	// maxSellLimit := util.MinInt64(sellLimit, sellOffer.Amount)
-	util.MinInt64(sellLimit, sellOffer.Amount)
+	sellLimit = util.MinInt64(sellLimit, offer.Amount)
 
-	return nil
-}
-
-// Get buying liability of provided offer
-func (of *Offer) GetBuyLiability(offer *ultpb.Offer) int64 {
-	return 0
-}
-
-// Get selling liability of provided offer
-func (of *Offer) GetSellLiability(offer *ultpb.Offer) int64 {
-	return 0
+	return sellLimit, buyLimit, nil
 }
 
 // Load selling and buying trust
@@ -217,6 +251,108 @@ func (of *Offer) validate() error {
 
 	if of.Amount == 0 && of.OfferID == "" {
 		return errors.New("offerID and amount are incompatible")
+	}
+
+	return nil
+}
+
+// Get buying liability of provided offer.
+func (of *Offer) getBuyingLiability() (int64, error) {
+	return 0, nil
+}
+
+// Get selling liability of provided offer.
+func (of *Offer) getSellingLiability() (int64, error) {
+	return 0, nil
+}
+
+// Update the liability of the account or the trust associated
+// with the offer.
+func (of *Offer) updateLiability(dt db.Tx, acquire bool) error {
+	// compute buying liability
+	bl, err := of.getBuyingLiability()
+	if err != nil {
+		return fmt.Errorf("get buying liability failed: %v", err)
+	}
+	if !acquire {
+		bl = -bl
+	}
+
+	if of.BuyAsset.AssetType == ultpb.AssetType_NATIVE {
+		acc, err := of.AM.GetAccount(dt, of.AccountID)
+		if err != nil {
+			return fmt.Errorf("get account failed: %v", err)
+		}
+
+		// update account buying liability
+		err = of.AM.UpdateLiability(acc, bl, true)
+		if err != nil {
+			return fmt.Errorf("update account buying liability failed: %v", err)
+		}
+
+		err = of.AM.SaveAccount(dt, acc)
+		if err != nil {
+			return fmt.Errorf("save account failed: %v", err)
+		}
+	} else {
+		trust, err := of.AM.GetTrust(dt, of.AccountID, of.BuyAsset)
+		if err != nil {
+			return fmt.Errorf("get trust failed: %v", err)
+		}
+
+		// update trust buying liability
+		err = of.AM.UpdateTrustLiability(trust, bl, true)
+		if err != nil {
+			return fmt.Errorf("update trust buying liability failed: %v", err)
+		}
+
+		err = of.AM.SaveTrust(dt, trust)
+		if err != nil {
+			return fmt.Errorf("save account failed: %v", err)
+		}
+	}
+
+	// compute selling liability
+	sl, err := of.getSellingLiability()
+	if err != nil {
+		return fmt.Errorf("get selling liability failed: %v", err)
+	}
+	if acquire {
+		sl = -sl
+	}
+
+	if of.SellAsset.AssetType == ultpb.AssetType_NATIVE {
+		acc, err := of.AM.GetAccount(dt, of.AccountID)
+		if err != nil {
+			return fmt.Errorf("get account failed: %v", err)
+		}
+
+		// update account buying liability
+		err = of.AM.UpdateLiability(acc, sl, false)
+		if err != nil {
+			return fmt.Errorf("update account buying liability failed: %v", err)
+		}
+
+		err = of.AM.SaveAccount(dt, acc)
+		if err != nil {
+			return fmt.Errorf("save account failed: %v", err)
+		}
+	} else {
+		trust, err := of.AM.GetTrust(dt, of.AccountID, of.BuyAsset)
+		if err != nil {
+			return fmt.Errorf("get trust failed: %v", err)
+		}
+
+		// update trust buying liability
+		err = of.AM.UpdateTrustLiability(trust, sl, true)
+		if err != nil {
+			return fmt.Errorf("update trust buying liability failed: %v", err)
+		}
+
+		err = of.AM.SaveTrust(dt, trust)
+		if err != nil {
+			return fmt.Errorf("save account failed: %v", err)
+		}
 	}
 
 	return nil
